docs: 规范化组件源码命名
